+++ src/client/index.ts
@@ -2,7 +2,7 @@ import { installClientProbes } from './events';
 import { installClientNativeBridge } from './native';
 
 const RESOURCE_NAME = GetCurrentResourceName();
-const VERSION = '0.2.0';
+const VERSION = '0.3.0';
 
 installClientProbes();
 installClientNativeBridge();